코드리뷰 수정반영
