+++ src/pages/seyeonPark/Main/MainComponent/AsideStory.js
@@ -13,7 +13,7 @@ class AsideStory extends React.Component {
             <img
               className="sideprofileimge2"
               alt="otherUserImg"
-              src="../../images/seyeonPark/profile2.jpeg"
+              src="/images/seyeonPark/profile2.jpeg"
             />
             <div className="userInfo">
               <p>
@@ -28,7 +28,7 @@ class AsideStory extends React.Component {
             <img
               className="sideprofileimge2"
               alt="otherUserImg"
-              src="../../images/seyeonPark/profile3.jpeg"
+              src="/images/seyeonPark/profile3.jpeg"
             />
             <div className="userInfo">
               <p>
@@ -43,7 +43,7 @@ class AsideStory extends React.Component {
             <img
               className="sideprofileimge2"
               alt="otherUserImg"
-              src="../../images/seyeonPark/profile4.jpeg"
+              src="/images/seyeonPark/profile4.jpeg"
             />
             <div className="userInfo">
               <p>
@@ -58,7 +58,7 @@ class AsideStory extends React.Component {
             <img
               className="sideprofileimge2"
               alt="otherUserImg"
-              src="../../images/seyeonPark/profile4.jpeg"
+              src="/images/seyeonPark/profile4.jpeg"
             />
             <div className="userInfo">
               <p>
